+++ core/src/main/java/ai/timefold/solver/core/impl/score/director/AbstractScoreDirectorFactory.java
@@ -1,11 +1,11 @@
 package ai.timefold.solver.core.impl.score.director;
 
-import java.util.function.Consumer;
-
 import ai.timefold.solver.core.api.domain.solution.PlanningSolution;
 import ai.timefold.solver.core.api.score.Score;
+import ai.timefold.solver.core.api.score.director.ScoreDirector;
 import ai.timefold.solver.core.config.solver.EnvironmentMode;
 import ai.timefold.solver.core.impl.domain.solution.descriptor.SolutionDescriptor;
+import ai.timefold.solver.core.impl.domain.variable.descriptor.BasicVariableDescriptor;
 import ai.timefold.solver.core.impl.domain.variable.descriptor.ListVariableDescriptor;
 import ai.timefold.solver.core.impl.score.constraint.ConstraintMatchPolicy;
 import ai.timefold.solver.core.impl.score.definition.ScoreDefinition;
@@ -109,8 +109,25 @@ public abstract class AbstractScoreDirectorFactory<Solution_, Score_ extends Sco
         }
     }
 
-    public Consumer<Object> getEntityValidator(InnerScoreDirector<Solution_, Score_> scoreDirector) {
-        return new EntityValidator<>(scoreDirector);
+    public void validateEntity(ScoreDirector<Solution_> scoreDirector, Object entity) {
+        if (listVariableDescriptor == null) { // Only basic variables.
+            var entityDescriptor = solutionDescriptor.findEntityDescriptorOrFail(entity.getClass());
+            if (entityDescriptor.isMovable(scoreDirector.getWorkingSolution(), entity)) {
+                return;
+            }
+            for (var variableDescriptor : entityDescriptor.getGenuineVariableDescriptorList()) {
+                var basicVariableDescriptor = (BasicVariableDescriptor<Solution_>) variableDescriptor;
+                if (basicVariableDescriptor.allowsUnassigned()) {
+                    continue;
+                }
+                var value = basicVariableDescriptor.getValue(entity);
+                if (value == null) {
+                    throw new IllegalStateException(
+                            "The entity (%s) has a variable (%s) pinned to null, even though unassigned values are not allowed."
+                                    .formatted(entity, basicVariableDescriptor.getVariableName()));
+                }
+            }
+        }
     }
 
 }
