+++ core/src/main/java/ai/timefold/solver/core/impl/solver/DefaultSolver.java
@@ -183,6 +183,7 @@ public class DefaultSolver<Solution_> extends AbstractSolver<Solution_> {
         solverScope.setInitialSolution(Objects.requireNonNull(problem, "The problem must not be null."));
         solverScope.setSolver(this);
         outerSolvingStarted(solverScope);
+
         var restartSolver = true;
         while (restartSolver) {
             var sample = solveLengthTimer.start();
@@ -223,6 +224,10 @@ public class DefaultSolver<Solution_> extends AbstractSolver<Solution_> {
         var startingSolverCount = solverScope.getStartingSolverCount() + 1;
         solverScope.setStartingSolverCount(startingSolverCount);
         registerSolverSpecificMetrics();
+
+        // Update the best solution, since problem's shadows and score were updated
+        bestSolutionRecaller.updateBestSolutionAndFireIfInitialized(solverScope);
+
         logger.info("Solving {}: time spent ({}), best score ({}), "
                 + "environment mode ({}), move thread count ({}), random ({}).",
                 (startingSolverCount == 1 ? "started" : "restarted"),
