+++ core/src/test/java/ai/timefold/solver/core/api/solver/SolutionManagerTest.java
@@ -9,7 +9,6 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.function.Function;
 
-import ai.timefold.solver.core.api.domain.variable.InverseRelationShadowVariable;
 import ai.timefold.solver.core.api.score.Score;
 import ai.timefold.solver.core.api.score.buildin.hardsoft.HardSoftScore;
 import ai.timefold.solver.core.api.score.buildin.simple.SimpleScore;
@@ -341,181 +340,6 @@ public class SolutionManagerTest {
         var score = solutionManager.update(solution, SolutionUpdatePolicy.UPDATE_SCORE_ONLY);
         assertThat(score).isEqualTo(HardSoftScore.ofHard(-4));
         assertThat(solution.getScore()).isEqualTo(HardSoftScore.ofHard(-4));
-    }
-
-    @ParameterizedTest
-    @EnumSource(SolutionManagerSource.class)
-    void updateOnlyScoreFailsIfShadowVariablesInconsistentIsNull(SolutionManagerSource SolutionManagerSource) {
-        var solution = new TestdataConcurrentSolution();
-        var e1 = new TestdataConcurrentEntity("e1");
-        var e2 = new TestdataConcurrentEntity("e2");
-
-        var a1 = new TestdataConcurrentValue("a1");
-        var a2 = new TestdataConcurrentValue("a2");
-        var b1 = new TestdataConcurrentValue("b1");
-        var b2 = new TestdataConcurrentValue("b2");
-
-        var groupA = List.of(a1, a2);
-        var groupB = List.of(b1, b2);
-
-        var entities = List.of(e1, e2);
-        var values = List.of(a1, a2, b1, b2);
-
-        a1.setConcurrentValueGroup(groupA);
-        a2.setConcurrentValueGroup(groupA);
-
-        b1.setConcurrentValueGroup(groupB);
-        b2.setConcurrentValueGroup(groupB);
-
-        e1.setValues(List.of(a1, b1));
-        e2.setValues(List.of(b2, a2));
-
-        b1.setPreviousValue(a1);
-        a2.setPreviousValue(b2);
-
-        a1.setNextValue(b1);
-        b2.setNextValue(a2);
-
-        a1.setIndex(0);
-        a2.setIndex(1);
-
-        b1.setIndex(1);
-        b2.setIndex(0);
-
-        a1.setEntity(e1);
-        b1.setEntity(e1);
-        a2.setEntity(e2);
-        b2.setEntity(e2);
-
-        a1.setInconsistent(true);
-        a2.setInconsistent(null);
-        b1.setInconsistent(true);
-        b2.setInconsistent(true);
-
-        solution.setEntities(entities);
-        solution.setValues(values);
-
-        var solutionManager = SolutionManagerSource.createSolutionManager(SOLVER_FACTORY_DECLARATIVE_SHADOW);
-        assertThat(solutionManager).isNotNull();
-        assertThatCode(() -> {
-            solutionManager.update(solution, SolutionUpdatePolicy.UPDATE_SCORE_ONLY);
-        }).hasMessageContainingAll(
-                "Shadow variables update is disabled",
-                "but the entity (e2 -> b2 -> a2)",
-                "has a null @ShadowVariablesInconsistent annotated field (isInconsistent)");
-    }
-
-    @ParameterizedTest
-    @EnumSource(SolutionManagerSource.class)
-    void updateOnlyScoreFailsIfInverseCollectionNull(SolutionManagerSource SolutionManagerSource) {
-        var solution = new TestdataInverseRelationSolution("solution");
-        var e1 = new TestdataInverseRelationEntity("e1");
-        var e2 = new TestdataInverseRelationEntity("e2");
-        var e3 = new TestdataInverseRelationEntity("e3");
-
-        var v1 = new TestdataInverseRelationValue("v1");
-        var v2 = new TestdataInverseRelationValue("v2");
-
-        v1.setEntities(null);
-        v2.setEntities(null);
-
-        solution.setEntityList(List.of(e1, e2, e3));
-        solution.setValueList(List.of(v1, v2));
-
-        e1.setValue(v1);
-        e2.setValue(v1);
-        e3.setValue(v2);
-
-        assertSoftly(softly -> {
-            softly.assertThat(solution.getScore()).isNull();
-            softly.assertThat(solution.getValueList().get(0).getEntities()).isNull();
-            softly.assertThat(solution.getValueList().get(1).getEntities()).isNull();
-        });
-
-        var solutionManager = SolutionManagerSource.createSolutionManager(SOLVER_FACTORY_INVERSE_RELATION);
-        assertThat(solutionManager).isNotNull();
-        assertThatCode(() -> {
-            solutionManager.update(solution, SolutionUpdatePolicy.UPDATE_SCORE_ONLY);
-        }).hasMessageContainingAll(
-                "The entity (v1)",
-                "has a collection @" + InverseRelationShadowVariable.class.getSimpleName(),
-                "that is null");
-    }
-
-    @ParameterizedTest
-    @EnumSource(SolutionManagerSource.class)
-    void updateOnlyScoreFailsIfInverseCollectionMissingElements(SolutionManagerSource SolutionManagerSource) {
-        var solution = new TestdataInverseRelationSolution("solution");
-        var e1 = new TestdataInverseRelationEntity("e1");
-        var e2 = new TestdataInverseRelationEntity("e2");
-        var e3 = new TestdataInverseRelationEntity("e3");
-
-        var v1 = new TestdataInverseRelationValue("v1");
-        var v2 = new TestdataInverseRelationValue("v2");
-
-        v1.setEntities(List.of(e1));
-        v2.setEntities(List.of(e3));
-
-        solution.setEntityList(List.of(e1, e2, e3));
-        solution.setValueList(List.of(v1, v2));
-
-        e1.setValue(v1);
-        e2.setValue(v1);
-        e3.setValue(v2);
-
-        assertSoftly(softly -> {
-            softly.assertThat(solution.getScore()).isNull();
-            softly.assertThat(solution.getValueList().get(0).getEntities()).containsExactly(e1);
-            softly.assertThat(solution.getValueList().get(1).getEntities()).containsExactly(e3);
-        });
-
-        var solutionManager = SolutionManagerSource.createSolutionManager(SOLVER_FACTORY_INVERSE_RELATION);
-        assertThat(solutionManager).isNotNull();
-        assertThatCode(() -> {
-            solutionManager.update(solution, SolutionUpdatePolicy.UPDATE_SCORE_ONLY);
-        }).hasMessageContainingAll(
-                "The entity (v1)",
-                "has a collection @" + InverseRelationShadowVariable.class.getSimpleName(),
-                "([e1])",
-                "that does not match the actual inverse set ([e1, e2])");
-    }
-
-    @ParameterizedTest
-    @EnumSource(SolutionManagerSource.class)
-    void updateOnlyScoreFailsIfInverseCollectionHasExtraElements(SolutionManagerSource SolutionManagerSource) {
-        var solution = new TestdataInverseRelationSolution("solution");
-        var e1 = new TestdataInverseRelationEntity("e1");
-        var e2 = new TestdataInverseRelationEntity("e2");
-        var e3 = new TestdataInverseRelationEntity("e3");
-
-        var v1 = new TestdataInverseRelationValue("v1");
-        var v2 = new TestdataInverseRelationValue("v2");
-
-        v1.setEntities(List.of(e1, e2));
-        v2.setEntities(List.of(e3, e1));
-
-        solution.setEntityList(List.of(e1, e2, e3));
-        solution.setValueList(List.of(v1, v2));
-
-        e1.setValue(v1);
-        e2.setValue(v1);
-        e3.setValue(v2);
-
-        assertSoftly(softly -> {
-            softly.assertThat(solution.getScore()).isNull();
-            softly.assertThat(solution.getValueList().get(0).getEntities()).containsExactly(e1, e2);
-            softly.assertThat(solution.getValueList().get(1).getEntities()).containsExactly(e3, e1);
-        });
-
-        var solutionManager = SolutionManagerSource.createSolutionManager(SOLVER_FACTORY_INVERSE_RELATION);
-        assertThat(solutionManager).isNotNull();
-        assertThatCode(() -> {
-            solutionManager.update(solution, SolutionUpdatePolicy.UPDATE_SCORE_ONLY);
-        }).hasMessageContainingAll(
-                "The entity (v2)",
-                "has a collection @" + InverseRelationShadowVariable.class.getSimpleName(),
-                "([e3, e1])",
-                "that does not match the actual inverse set ([e3])");
     }
 
     @ParameterizedTest
